Fix function inner members restore
